requirements.txt
openai
streamlit
transformers
flask
requests
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
transformers
torch
huggingface-hub
streamlit_feedback